+++ Sources/Views/Common/SizeReader.swift
@@ -1,0 +1,17 @@
+import SwiftUI
+
+struct SizeReader: View {
+    @Binding var size: CGSize
+
+    var body: some View {
+        GeometryReader { geometry in
+            Color.clear
+                .onAppear {
+                    size = geometry.size
+                }
+                .onChange(of: geometry.size) { _, newValue in
+                    size = newValue
+                }
+        }
+    }
+}